Drop the iCloud key-value mirror from Apple client settings
Settings no longer read from or write to NSUbiquitousKeyValueStore, not
even as a read fallback. UserDefaultsPreferenceStore (plain local
UserDefaults) replaces UbiquitousPreferenceStore; cross-device sync is
solely the per-Cognito-user server row via PreferencesSyncCoordinator.

The iCloud layer was the wrong identity anchor (Apple ID, one shared
store for every Cabalmail account on it) and was dead weight in
practice: neither app carries the ubiquity-kvstore entitlement, so the
mirror never actually synced — which also means there is nothing in
iCloud to scrub. The leaked legacy values live only in local
UserDefaults, which activate() already purges.
