+++ changelog.d/apple-per-account-preferences.fixed.md
@@ -7,4 +7,7 @@
   local settings scope, activated at sign-in before the server copy is
   pulled, and the old device-wide values are deleted on first sign-in
   after updating (settings re-sync from the server, or reset to defaults
-  for accounts that never synced).
+  for accounts that never synced). The local cache is now plain
+  `UserDefaults` on this device only — the iCloud key-value mirror is
+  gone entirely, so nothing is read from or written to iCloud, even as a
+  fallback; cross-device sync rides the Cabalmail account.
